+++ pepselect-shipping-restrictions/tests/shipping-restrictions-contract.php
@@ -26,16 +26,16 @@ final class PepSelect_Test_Rate {
 pepselect_shipping_assert( PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'CA', '90210' ), 'California should pass.' );
 pepselect_shipping_assert( PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'AK', '99501' ), 'Alaska should pass.' );
 pepselect_shipping_assert( PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'HI', '96801' ), 'Hawaii should pass.' );
-pepselect_shipping_assert( PepSelect_Shipping_Restrictions::address_is_allowed( 'PR', '', '00901' ), 'Puerto Rico country and ZIP should pass.' );
+pepselect_shipping_assert( PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'PR', '00901' ), 'Puerto Rico state and ZIP should pass.' );
 pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'NY', '00926' ), 'Puerto Rico ZIP with New York selected should fail.' );
-pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'PR', '00901' ), 'Puerto Rico must be selected as the country.' );
-pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'PR', '', '10001' ), 'New York ZIP with Puerto Rico country selected should fail.' );
+pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'PR', '', '00901' ), 'Puerto Rico country code should not bypass the U.S. destination rules.' );
+pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'PR', '10001' ), 'New York ZIP with Puerto Rico selected should fail.' );
 pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'VI', '00802' ), 'U.S. Virgin Islands should fail.' );
 pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'HI', '96799' ), 'American Samoa should fail.' );
 pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'CA', '96201' ), 'Overseas military ZIP should fail.' );
 pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::address_is_allowed( 'US', 'CA', '96910' ), 'Pacific territory ZIP should fail.' );
 pepselect_shipping_assert( PepSelect_Shipping_Restrictions::destination_is_usps_only( 'US', 'AK', '99501' ), 'Alaska should be USPS-only.' );
-pepselect_shipping_assert( PepSelect_Shipping_Restrictions::destination_is_usps_only( 'PR', '', '00901' ), 'Puerto Rico should be USPS-only.' );
+pepselect_shipping_assert( PepSelect_Shipping_Restrictions::destination_is_usps_only( 'US', 'PR', '00901' ), 'Puerto Rico should be USPS-only.' );
 pepselect_shipping_assert( ! PepSelect_Shipping_Restrictions::destination_is_usps_only( 'US', 'HI', '96801' ), 'Hawaii should use returned live carrier options.' );
 
 $restriction = PepSelect_Shipping_Restrictions::instance();
@@ -49,7 +49,7 @@ pepselect_shipping_assert(
 	array( 'easyship:usps' ) === array_keys(
 		$restriction->filter_package_rates(
 			$rates,
-			array( 'destination' => array( 'country' => 'PR', 'state' => '', 'postcode' => '00901' ) )
+			array( 'destination' => array( 'country' => 'US', 'state' => 'PR', 'postcode' => '00901' ) )
 		)
 	),
 	'Puerto Rico should retain only USPS.'